+++ src/main/java/roomescape/global/error/BusinessException.java
@@ -1,4 +1,4 @@
-package roomescape.error;
+package roomescape.global.error;
 
 import org.springframework.http.HttpStatus;
