+++ src/main/java/roomescape/global/error/ErrorCode.java
@@ -1,4 +1,4 @@
-package roomescape.error;
+package roomescape.global.error;
 
 public enum ErrorCode {
     TIME_NOT_FOUND("예약 시간을 찾을 수 없습니다."),
@@ -13,9 +13,8 @@ public enum ErrorCode {
     INTERNAL_SERVER_ERROR("서버 오류가 발생했습니다."),
     RESERVATION_FORBIDDEN_REQUEST("예약에 대한 접근 권한이 없습니다."),
     RESERVATION_EXPIRED("이미 시간이 지난 예약은 수정 및 삭제할 수 없습니다."),
+    CONCURRENT_REQUEST("동시 요청으로 인해 처리하지 못했습니다. 잠시 후 시도해 주세요.")
     ;
-
-
 
     private final String message;
