+++ src/main/java/roomescape/reservation/exception/ForbiddenRequestException.java
@@ -1,8 +1,8 @@
 package roomescape.reservation.exception;
 
 import org.springframework.http.HttpStatus;
-import roomescape.error.BusinessException;
-import roomescape.error.ErrorCode;
+import roomescape.global.error.BusinessException;
+import roomescape.global.error.ErrorCode;
 
 public class ForbiddenRequestException extends BusinessException {
     public ForbiddenRequestException() {
