+++ src/main/java/roomescape/theme/exception/ThemeNotFoundException.java
@@ -2,8 +2,8 @@ package roomescape.theme.exception;
 
 import org.springframework.http.HttpStatus;
 
-import roomescape.error.BusinessException;
-import roomescape.error.ErrorCode;
+import roomescape.global.error.BusinessException;
+import roomescape.global.error.ErrorCode;
 
 public class ThemeNotFoundException extends BusinessException {
     public ThemeNotFoundException(Long id) {
